requirements.txt
# Core stack
pandas>=2.0
numpy>=1.23
scipy>=1.10
scikit-learn>=1.3
pyyaml>=6.0
joblib>=1.2
matplotlib>=3.7
python-dateutil>=2.8.2

# Parquet IO (required by scripts that read/write .parquet)
pyarrow>=14.0.1

# Project libraries used in code
fastf1>=3.3.0
requests>=2.31.0
requests-cache>=1.1.0
python-dotenv>=1.0.0

# add project dependencies here
# kafka-python==2.0.2
feast==0.18.0
lightgbm==3.3.2
mlflow>=2.10.0
wandb>=0.16.0

# Geospatial stack used by track/elevation scripts
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0

# code style & linting
black>=25.1.0
isort>=5.10.1
flake8>=7.3.0
pre-commit>=3.5.0

# tuning/exp mgmt
optuna>=3.0.0
